pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexus-alpha"
version = "3.0.0"
description = "Self-evolving crypto trading scaffold with modular signal, risk, execution, and world-model layers."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Proprietary"}

dependencies = [
    # === Classical ML ===
    "scikit-learn>=1.5.0",

    # === Data / Time Series ===
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "scipy>=1.13.0",
    "statsmodels>=0.14.0",

    # === API / Async ===
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "websockets>=13.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",

    # === Message Queue / Streaming ===
    "confluent-kafka>=2.4.0",
    "aiokafka>=0.10.0",

    # === Database ===
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "psycopg2-binary>=2.9.9",

    # === Observability ===
    "prometheus-client>=0.20.0",
    "structlog>=24.1.0",
    "opentelemetry-api>=1.24.0",
    "opentelemetry-sdk>=1.24.0",

    # === Configuration ===
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",

    # === Crypto Exchange Clients ===
    "ccxt>=4.3.0",              # 100+ exchanges, free public data
    "ccxt[async]>=4.3.0",       # ccxt.pro WebSocket streaming (free)

    # === NLP / LLM — Free Stack ===
    # Ollama client (OpenAI-compatible local inference)
    "openai>=1.30.0",           # Works as Ollama client via base_url override
    # Groq free-tier cloud fallback (6k req/day, Llama-3.3-70b free)
    "groq>=0.9.0",
    # FinBERT + local embeddings via HuggingFace
    "transformers>=4.41.0",
    "sentence-transformers>=3.0.0",
    # Free web intelligence (replaces OpenClaw paid backend)
    "crawl4ai>=0.3.0",
    # Free news sources
    "feedparser>=6.0.0",
    "newspaper4k>=0.9.0",       # newspaper3k maintained fork — Python 3.11+ compatible
    "praw>=7.7.0",              # Reddit API (60 req/min free)
    # Free alerting (replaces PagerDuty)
    "python-telegram-bot>=21.0",
    # Free vector DB client
    "qdrant-client>=1.9.0",
    # Job scheduling for intelligence agents
    "apscheduler>=3.10.0",
    # Free on-chain / sentiment data
    "pycoingecko>=3.1.0",

    # === Utilities ===
    "click>=8.1.0",
    "rich>=13.7.0",
    "tenacity>=8.3.0",
    "orjson>=3.10.0",

    # === Core ML / Reinforcement Learning (v6 Core) ===
    "torch>=2.3.0",
    "stable-baselines3>=2.3.0",
    "gymnasium>=0.29.0",
    "dowhy>=0.11.0",
    "lingam>=1.8.0",
    "lz4>=4.3.0",
]

[project.optional-dependencies]
research = [
    # === Deep Learning Extensions ===
    "pytorch-lightning>=2.2.0",
    "pytorch-forecasting>=1.1.0",
    # === Genetic Programming / Evolutionary ===
    "deap>=1.4.1",
    # === Classical ML boosters ===
    "xgboost>=2.0.0",
    "lightgbm>=4.3.0",
    # === Explainability / tuning / tracking ===
    "shap>=0.45.0",
    "optuna>=3.6.0",
    "mlflow>=2.12.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
]

[project.scripts]
nexus = "nexus_alpha.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["nexus_alpha*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "B017", "SLF001"]
"tests/integration/conftest.py" = ["S603", "S607"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "kafka_integration: requires docker compose kafka stack and confluent_kafka",
]
